(#79): detination하위로 폴더 빼기
