--- src/pages/map/destination/components/follow/index.ts
@@ -1,2 +1,0 @@
-export { default as FollowAlong } from './FollowAlong';
-export { default as SelectedFollow } from './SelectedFollow';
