+++ src/pages/map/destination/components/index.ts
@@ -1,4 +1,8 @@
-export { default as ResizingMap } from './ResizingMap';
 export { default as GeoDestinationMap } from './GeoDestinationMap';
+
+export { default as ResizingMap } from './ResizingMap';
 export { default as SelectTransportationFromGeoMap } from './SelectTransportationFromGeoMap';
 export { default as withDestination } from './withDestination';
+export { default as FollowAlong } from './FollowAlong';
+export { default as SelectedFollow } from './SelectedFollow';
+export { default as PrefetchMap } from './PrefetchMap';
