+++ src/pages/map/destination/lib/index.ts
@@ -1,1 +1,4 @@
 export { default as useMapController } from './useMapController';
+export { default as useAddressFromCoords } from './useAddressFromCoords';
+
+export { default as isSelectedOptions } from './isSelectedOptions';
