+++ src/pages/map/destination/pedestrian/index.ts
